Refactor CLI into modular components
- Move evaluation CLI commands into separate modules
- Create a modular directory structure in src/eval/cli
- Split into individual command files for better maintainability
- Remove duplicated code from main CLI file
- Improve type safety across the implementation
